Fix chem initialization bugs in MOZART and MOSAIC
Add initializations for several arrays in MOZART and MOSAIC in chem, in
the aerosol chemistry code, in the profile initialization for chemical
variables in real, and in the irr_diag reaction rate output.
These issues trigger crashes at runtime at the model initialization
stage when using debug compilation flags, and were preventing the use of
debug options for investigating other bugs.
In addition, when using gfortran, these arrays would be be initialized
with arbitrary values, then used in the chemistry, causing some aerosol
output to diverge to 0 or infinity.